+++ app/services/activity/roda_identifier_generator.rb
@@ -1,8 +1,9 @@
 class Activity
   class RodaIdentifierGenerator
-    def initialize(parent_activity:, extending_organisation:)
+    def initialize(parent_activity:, extending_organisation:, is_non_oda: false)
       @parent_activity = parent_activity
       @extending_organisation = extending_organisation
+      @is_non_oda = is_non_oda
     end
 
     def generate
@@ -16,11 +17,14 @@ class Activity
     private
 
     def programme_identifier
+      non_oda_prefix = @is_non_oda ? "NODA" : nil
+
       [
+        non_oda_prefix,
         parent_activity.roda_identifier,
         extending_organisation.beis_organisation_reference,
         component_identifier
-      ].join("-")
+      ].compact.join("-")
     end
 
     def project_identifier