+++ spec/services/forecast/import_spec.rb
@@ -31,25 +31,83 @@ RSpec.describe Forecast::Import do
   end
 
   describe "#imported_forecasts" do
-    let(:forecast) { double("forecast") }
+    context "a generic import" do
+      let(:forecast) { double("forecast") }
 
-    let :forecast_row do
-      {
-        "Activity RODA Identifier" => project.roda_identifier,
-        "FC 2020/21 FY Q3 (Oct, Nov, Dec)" => "200436",
-        "FC 2020/21 FY Q4 (Jan, Feb, Mar)" => "310793"
-      }
+      let :forecast_row do
+        {
+          "Activity RODA Identifier" => project.roda_identifier,
+          "FC 2020/21 FY Q3 (Oct, Nov, Dec)" => "200436",
+          "FC 2020/21 FY Q4 (Jan, Feb, Mar)" => "310793"
+        }
+      end
+
+      before do
+        forecast_history = instance_double(ForecastHistory, set_value: forecast, latest_entry: forecast)
+        allow(ForecastHistory).to receive(:new).and_return(forecast_history)
+      end
+
+      it "returns a list of forecasts" do
+        importer.import([forecast_row])
+
+        expect(importer.imported_forecasts).to eq([forecast, forecast])
+      end
     end
 
-    before do
-      forecast_history = instance_double(ForecastHistory, set_value: forecast)
-      allow(ForecastHistory).to receive(:new).and_return(forecast_history)
-    end
+    context "importing into a specific report" do
+      let(:selected_report) { latest_report }
+      let(:reporter) { nil }
 
-    it "returns a list of forecasts" do
-      importer.import([forecast_row])
+      context "when the report already has some original forecasts" do
+        let :forecast_row do
+          {
+            "Activity RODA Identifier" => project.roda_identifier,
+            "FC 2020/21 FY Q3 (Oct, Nov, Dec)" => "200436",
+            "FC 2020/21 FY Q4 (Jan, Feb, Mar)" => "310793",
+            "FC 2021/22 FY Q1 (Apr, May, Jun)" => "984150",
+            "FC 2021/22 FY Q2 (Jul, Aug, Sep)" => "206206"
+          }
+        end
 
-      expect(importer.imported_forecasts).to eq([forecast, forecast])
+        before do
+          Forecast::Import.new(uploader: reporter, report: selected_report).import([forecast_row])
+        end
+
+        context "when replacing some existing values with zero" do
+          let(:updated_forecast_row) do
+            {
+              "Activity RODA Identifier" => project.roda_identifier,
+              "FC 2020/21 FY Q3 (Oct, Nov, Dec)" => "0",
+              "FC 2020/21 FY Q4 (Jan, Feb, Mar)" => "0"
+            }
+          end
+
+          it "does not include the deleted forecasts" do
+            importer.import([updated_forecast_row])
+
+            expect(importer.imported_forecasts).to eq([nil, nil])
+          end
+        end
+
+        context "when re-uploading forecasts that are the same as the existing values" do
+          let(:updated_forecast_row) do
+            {
+              "Activity RODA Identifier" => project.roda_identifier,
+              "FC 2020/21 FY Q3 (Oct, Nov, Dec)" => "200436",
+              "FC 2020/21 FY Q4 (Jan, Feb, Mar)" => "310793"
+            }
+          end
+
+          it "includes the unchanged values" do
+            importer.import([updated_forecast_row])
+
+            expect(importer.imported_forecasts.map { |f| [f.financial_quarter, f.financial_year, f.value] }).to eq([
+              [3, 2020, 200_436.0],
+              [4, 2020, 310_793]
+            ])
+          end
+        end
+      end
     end
   end