+++ CHANGELOG.md
@@ -1195,6 +1195,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 ## [release-130] 2023-02-13
 
 - Update Ruby to 3.0.5
+- Prefix imported non-ODA programmes "NODA-"
 
 ## [unreleased]