+++ CHANGELOG.md
@@ -1195,12 +1195,13 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 ## [release-130] 2023-02-13
 
 - Update Ruby to 3.0.5
-- Prefix imported non-ODA programmes "NODA-"
 
 ## [unreleased]
 
 - Remove heading for unpopulated organisation column in organisation reports table for BEIS users
 - Prevent child ODA activities being created on non-ODA parents and vice-versa via the bulk upload
+- Prefix imported non-ODA programmes "NODA-"
+- Prevent editing ODA type via the UI
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-130...HEAD
 [release-130]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-129...release-130