+++ CHANGELOG.md
@@ -1202,6 +1202,9 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 - Prevent child ODA activities being created on non-ODA parents and vice-versa via the bulk upload
 - Prefix imported non-ODA programmes "NODA-"
 - Prevent editing ODA type via the UI
+- Fix the display of successfully uploaded forecasts for the cases when:
+  - the original forecast is deleted
+  - the imported values were already in the report
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-130...HEAD
 [release-130]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-129...release-130