+++ CHANGELOG.md
@@ -1209,6 +1209,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 - Remove legacy fields from ISPF reports.
 - Only generate IATI identifiers for ODA activities
 - Remove reference to truncating data published to Statistics in International Development for non-ODA activities
+- Hide IATI identifier and "Include in IATI XML export?" fields on activity details tab for non-ODA activities
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-130...HEAD
 [release-130]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-129...release-130